+++ src/base/SubscriptionGateway.php
@@ -495,10 +495,10 @@ abstract class SubscriptionGateway extends Gateway
         switch ($data['type']) {
             case 'payment_method.detached':
                 $this->handlePaymentMethodDetached($data);
-                // no break
+            // no break
             case 'charge.refund.updated':
                 $this->handleRefundUpdated($data);
-                // no break
+            // no break
             case 'plan.deleted':
             case 'plan.updated':
                 $this->handlePlanEvent($data);
@@ -684,14 +684,20 @@ abstract class SubscriptionGateway extends Gateway
         $this->configureStripeClient();
         $planService = Commerce::getInstance()->getPlans();
 
-        if ($data['type'] == 'plan.deleted') {
-            $plan = $planService->getPlanByReference($data['data']['object']['id']);
-
-            if ($plan) {
-                $planService->archivePlanById($plan->id);
-                Craft::warning($plan->name . ' was archived because the corresponding plan was deleted on Stripe. (event "' . $data['id'] . '")', 'stripe');
-            }
+        $plan = $planService->getPlanByReference($data['data']['object']['id']);
+        
+        if (!$plan) {
+            throw new InvalidConfigException('Plan with the reference “' . $data['data']['object']['id'] . '” not found when processing webhook ' . $data['id']);
         }
+
+        if ($data['type'] == 'plan.deleted') {
+            $planService->archivePlanById($plan->id);
+            Craft::warning($plan->name . ' was archived because the corresponding plan was deleted on Stripe. (event "' . $data['id'] . '")', 'stripe');
+        } else {
+            $plan->planData = Json::encode($data['data']['object']);
+            $planService->savePlan($plan);
+        }
+
     }
 
     /**